+++ 2022/07/Directory.cs
@@ -5,7 +5,7 @@ public class Directory
     private string Id { get; }
     public HashSet<Directory> SubDirectories { get; } = new();
     public List<File> Files { get; } = new();
-    public long Size { get; private set; }
+    public int Size { get; private set; }
 
     public Directory(string id)
     {
@@ -17,7 +17,7 @@ public class Directory
         Size = GetSize();
     }
 
-    private long GetSize()
+    private int GetSize()
     {
         return Files.Sum(f => f.Size) + SubDirectories.Sum(d => d.GetSize());
     }
